work more on the frontend
