+++ milestone2/frontend/Dockerfile
@@ -12,8 +12,12 @@ RUN rm -rf ./components
 RUN rm -rf ./assets
 RUN mkdir -p components
 RUN mkdir -p assets
+RUN mkdir -p assets/css
+RUN mkdir -p assets/img
+
 COPY ./components ./components
-COPY ./assets ./assets
+COPY ./assets/css ./assets/css
+COPY ./assets/img ./assets/img
 
 # Install dependencies
 RUN npm install --production
