+++ milestone2/frontend/assets/css/map.css
@@ -1,0 +1,55 @@
+/* Placeholder style for the plot */
+#myPlot {
+    width: 100%;
+    height: 300px;
+    background-color: lightgray;
+}
+
+#map-canvas { 
+    height: 650px; 
+    width: 700px;    
+}
+
+.placement-heatmap {
+    width: 700px; /* Set the width of the container */
+    margin: 50px auto; /* Center the container horizontally */
+    padding: 20px;
+    background-color: #fff; /* Background color of the container */
+    border-radius: 10px; /* Rounded corners */
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2); /* Shadow effect */
+}
+
+.map-container {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start; /* Align items to the top */
+    max-width: 1200px; /* Adjust the maximum width as needed */
+    margin: 0 auto; /* Center the container horizontally */
+    padding: 20px;
+}
+
+.post-its.left {
+    display: flex;
+    flex-direction: column;
+    margin-top: 200px; /* Adjust the upper margin for the left post-its */
+    margin-right: 100px;
+    margin-left: -250px;
+}
+
+.post-its.right {
+    display: flex;
+    flex-direction: column;
+    margin-top: 300px;
+    margin-left: 100px;
+}
+
+.post-it {
+    width: 200px;
+    padding: 15px;
+    background-color: #ffd000; /* Yellow background */
+    border: 2px solid #ffbf00; /* Yellow border */
+    border-radius: 10px;
+    margin: 100px;
+    box-shadow: 3px 3px 5px rgba(0, 0, 0, 0.2); /* Shadow effect */
+    margin-bottom: 20px;
+}
