+++ milestone2/frontend/components/predictions.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <link rel="stylesheet" href="../assets/css/style.css">
+    <link rel="stylesheet" href="../assets/css/prediction.css">
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" integrity="sha512-xodZBNTC5n17Xt2atTPuE1HxjVMSvLVW9ocqUKLsCC5CXdbqCmblAshOMAS6/keqq/sMZMZ19scR4PsZChSR7A==" crossorigin=""/>
+
+    <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js" integrity="sha512-XQoYMqMTK8LvdxXYG3nZ448hOEQiglfqkJs1NOQV44cWnUrBc8PkAOcXy20w0vlaXaVUearIOBhiXZ5V3ynxwA==" crossorigin=""></script>
+</head>
+
+<nav class="menu">
+    <div class="menu">
+        <ul class="menu-bar">
+            <li class="toolbar-title">Ride Sharing Impact</li>
+            <li class="toolbar-separator"></li>
+            <li><a class="nav-link" href="../index.html"><img src="../assets/img/home.png" alt="Home" aria-hidden="true" class="menu-icon">Home</a></li>
+            <li><a class="nav-link" href="heatmaps.html"><img src="../assets/img/heatmap.png" alt="Heat Maps" aria-hidden="true" class="menu-icon">HeatMaps</a></li>
+            <li><a class="nav-link" href="timeseries.html"><img src="../assets/img/timeseries.png" alt="Time Series" aria-hidden="true" class="menu-icon">TimeSeries</a></li>
+            <li><a class="nav-link" href="predictions.html"><img src="../assets/img/predictions.png" alt="Predictions" aria-hidden="true" class="menu-icon">Predictions</a></li>
+        </ul>
+    </div>
+</nav>
+<hr style="height:2px;border-width:0;color:rgb(176, 176, 176);background-color:rgb(176, 176, 176)">
+<body>
+    <div id="book">
+        <div class="page">
+            <div id="map"></div>
+        </div>
+        <div class="page">
+            <div id="prediction-box">
+                <h2>Prediction</h2>
+                <p>Selected coordinates:</p>
+                <div id="coordinates"></div>
+                <p>Accident prediction:</p>
+                <div id="prediction"></div>
+            </div>
+        </div>
+    </div>
+    <script src="predictions.js"></script>
+</body>
+</html>
