+++ milestone2/frontend/components/timeseries.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <link rel="stylesheet" href="../assets/css/style.css">
+</head>
+
+<nav class="menu">
+    <div class="menu">
+        <ul class="menu-bar">
+            <li class="toolbar-title">Ride Sharing Impact</li>
+            <li class="toolbar-separator"></li>
+            <li><a class="nav-link" href="../index.html"><img src="../assets/img/home.png" alt="Home" aria-hidden="true" class="menu-icon">Home</a></li>
+            <li><a class="nav-link" href="heatmaps.html"><img src="../assets/img/heatmap.png" alt="Heat Maps" aria-hidden="true" class="menu-icon">HeatMaps</a></li>
+            <li><a class="nav-link" href="timeseries.html"><img src="../assets/img/timeseries.png" alt="Time Series" aria-hidden="true" class="menu-icon">TimeSeries</a></li>
+            <li><a class="nav-link" href="predictions.html"><img src="../assets/img/predictions.png" alt="Predictions" aria-hidden="true" class="menu-icon">Predictions</a></li>
+        </ul>
+    </div>
+</nav>
+<hr style="height:2px;border-width:0;color:rgb(176, 176, 176);background-color:rgb(176, 176, 176)">
+</html>
