+++ milestone2/frontend/index.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://unpkg.com/leaflet/dist/leaflet.css" />
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/noUiSlider/15.2.0/nouislider.min.css" />
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/flatpickr/dist/flatpickr.min.css">
-    <link rel="stylesheet" href="./assets/slider.css">
+    <link rel="stylesheet" href="./assets/css/slider.css">
+    <link rel="stylesheet" href="./assets/css/style.css">
+    <link rel="stylesheet" href="./assets/css/vehicles.css">
 
     <!-- Dependencies -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.1/Chart.js"></script>
@@ -18,71 +18,38 @@
     <script src="https://d3js.org/d3.v7.min.js"></script>
     
     <title>Ride Sharing Impact</title>
-    <style>
-        /* Basic styling for the header */
-        header {
-            background-color: #a0841f;
-            color: #000000;
-            padding: 10px 0;
-            text-align: center;
-        }
-
-        /* Styling for the title */
-        h1 {
-            margin: 0;
-            font-size: 24px;
-        }
-
-         .quote {
-            font-style: italic;
-            margin: 20px auto; /* Center the quote horizontally */
-            padding: 20px; /* Add padding to create space around the quote */
-            max-width: 600px; /* Set maximum width to control text wrapping */
-            background-color: #f9f9f9; /* Background color for the quote box */
-            border: 1px solid #ccc; /* Border for the quote box */
-            border-radius: 8px; /* Rounded corners for the quote box */
-            text-align: center; /* Center-align the text */
-            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-        }
-
-         /* Container for the main content */
-         .container {
-            display: flex;
-        }
-
-        /* Style for the left column */
-        .left-column {
-            flex: 1;
-            padding: 20px;
-        }
-
-        /* Style for the right column */
-        .right-column {
-            flex: 1;
-            padding: 20px;
-        }
-
-        /* Placeholder style for the plot */
-        #myPlot {
-            width: 100%;
-            height: 300px;
-            background-color: lightgray;
-        }
-
-        #map-canvas { 
-            height: 650px; 
-            width: 700px;    
-        }
-    </style>
 </head>
+
+<nav class="menu">
+    <div class="menu">
+        <ul class="menu-bar">
+            <li class="toolbar-title">Ride Sharing Impact</li>
+            <li class="toolbar-separator"></li>
+            <li><a class="nav-link" href="index.html"><img src="./assets/img/home.png" alt="Home" aria-hidden="true" class="menu-icon">Home</a></li>
+            <li><a class="nav-link" href="./components/heatmaps.html"><img src="./assets/img/heatmap.png" alt="Heat Maps" aria-hidden="true" class="menu-icon">HeatMaps</a></li>
+            <li><a class="nav-link" href="./components/timeseries.html"><img src="./assets/img/timeseries.png" alt="Time Series" aria-hidden="true" class="menu-icon">TimeSeries</a></li>
+            <li><a class="nav-link" href="./components/predictions.html"><img src="./assets/img/predictions.png" alt="Predictions" aria-hidden="true" class="menu-icon">Predictions</a></li>
+        </ul>
+    </div>
+</nav>
+<hr style="height:2px;border-width:0;color:rgb(176, 176, 176);background-color:rgb(176, 176, 176)">
 <body>
-    <header>
-        <h1>Ride Sharing Impact</h1>
-    </header>
+
+    <section class="vehicle-container">
+        <!-- Taxi Icon -->
+        <div class="vehicle">
+          <img src="./assets/img/taxi_hailing.jpeg" alt="Taxi" class="vehicle-icon">
+        </div>
+        <!-- Bike Icon -->
+        <div class="vehicle">
+          <img src="./assets/img/bike.jpg" alt="Bike" class="vehicle-icon">
+        </div>
+    </section>
+
     <div class="quote">
         <p>"Four wheels move the body; two wheels move the soul. Life is short, buy the motorcycle, have a ride, live your dreams. You don't stop riding when you get old; you get old when you stop riding."</p>
     </div>
-    <div class="container">
+    <!-- <div class="container">
         <div class="left-column">
             <p>...Magic Spells Potions and Secrets...</p>
         </div>
@@ -91,25 +58,7 @@
             <canvas id="myChart"></canvas>
         </div>
     </div>
-    
-    
-    <div>
-        <label for="resource">Select Resource:</label>
-        <select id="resource">
-            <option value="taxis">Taxis</option>
-            <option value="bikes">Bikes</option>
-            <option value="accidents">Accidents</option>
-        </select>
-    </div>
-    
-    <div id="map-canvas"></div> 
-
-    <div id="slider-container">
-        <svg id="slider">
-            <text x="50%" y="100%" text-anchor="middle" alignment-baseline="middle">Choose timeframe</text>
-        </svg>
-    </div>
-    
+     -->
     <script type="module" src="main.js"></script>
 </body>
 </html>
