+++ milestone2/frontend/package.json
@@ -12,6 +12,14 @@
   "plugins": [
     "@babel/plugin-transform-modules-commonjs"
   ],
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^4.2.0"
+  },
   "author": "OutOfTouch",
   "license": "ISC",
   "main": "main.js",
